Fixed TestSuite, Added GameController and Added IGameViewOutputController
Fixed issues with Test Runner to now allow PlayMode and EditMode Tests.
Added GameController which takes in input from the View, interputs it for the PlayerManager and outputs results to the OutputViewController implementing IGameVIewOutputController.
Added IGameViewOutputController for implementation by the ViewController that takes is interested in the update of Player and GameState.