Refactored PlayerData and IGameViewOutputController, Added EnergyData and Implemented GameController
Refactored PlayerData and IGameViewOutputController to use Players enum for ID purposes.
Added EnergyData to mirror PlayerData's use of being used to pass data from GameController to GameView.
Implemented GameController to damage a given player and update the game view with new PlayerData, update GameView that a player has won and added a helper function that converts PlayerModel to PlayerData.