Player health and energy sync'd between server and clients
Create LightWarNetworkManager to override the adding of player to being able to connect the created PlayerViews to the PayerOutputs and connect the GameManager to the PlayerVIews.
Created a Network main to create the GameManager, PlayerOutputController and connect them to the PlayerViews.
NetworkPlayerOutputController does its regualar passing around of data from GameManager to NetworkPlayerViews but is set to only work on the Server.
NetworkPlayerVIew networks the health and energy on Server and Client.
NetworkMain object added to the Scene to facilitate all of the creation.