Created EnergyManager and IEnergyLoader to handle the loading and storing of different types of energies in a game
Created the IEnergyLoader to be an interface between the different types of storing and reading Energy types. Acts as solidified API for EnergyManger to use.
Created the EnergyManager to create from a IEnergyLoader and store all pertanent information about what types of energies will be used in a game. EnergyManager stores the EnergyData into two portions the EnergyObjectData which stores all data pertaining to its model and EnergyType data which stores redundant data to allow for quick display of types for Player to choose from.
Added EnergyManagerTests to test its ability to store and connect the EnergyData and its view data used to created visual for what energies a player can spawn;