Switched from EnergyManager and EnergyModels to EnergySpawner and EnergyType
Switched EnergyManager to BaseEnergySpawner. BaseEnergySpawner does everything EnergyManager does as well as provides an interface to spawn Energies based on the types it gives.
Switched EnergyModel to EnergyType. Basic name change to better illistrate what it does. Used to define types for the BaseEnergySpawner to create and display with.
Removed EnergyModel tests.
Modified EnergyManagerTests to represent BaseEneregySpawner.