Rearranged the folder structure
Added Game Logic, Single Player, and Multiplayer folders in both Code and Prefabs folders.
Game Logic consists of base logic which both Single and Multiplayer use to make the game go.
Single Player consists of code pertaining to the views, controllers and main of a single player game.
Multiplayer consists of the code pertaininf to the views, controllers, main and networking specific code for the setup of a multiplayer game.