Energies now spawn on owning player and move towards enemy player in respects to the clients
Network Energy Spawner spawns Energy and gives it a owning player id and initializes its speed (hard coded at the moment).
Network Energy View decides on where it starts and where it ends based on what client its on and who owns it.
Network Energy View moved based on a percentage which is syned between clients to allow for them to move in opposing directions on clients in respect to the client.
Network Player VIew now syncs its Id with the server and call the Network Energy Spawner when clicked to spawn an energy.
Network main now initializes the Eneregy Spawner and passes it to the views.
Created Constants class to hold global constants for the use of positions of local and non local player.