NetworkPlayerInput now decrements energy when casting energy
NetworkPlayerInput now checks with GameManager if it has enough energy to cast and energy and if so spawns an energy.
GameManager now provides a check to see if a player has enough energy to cast an energy and if so decrements the energy of said player and updates its visual.
PlayerController now provides access to decrementing a players energy by a given amount.