Removed PlayerIds enum
Removed the use of PlayerIds enum from the code base as it created more complexity than needed. Now all using and refering of players are with and Integer id corresponding to their position in respective arrays.
Any checks on either player can use the Constants file to grab the constant respective id for each player and allow for more descriptive code than just 0 or 1.