Added Game Logic functionality to Load and store spawn points
Created SpawnPoint model to store the location, owner and siblings of each spawn point in the game.
Created BaseSpawnPointSpawner to call upon the ISpawnPointLoader to load in and store all the information need for the Spawn Points in the game.
Created the SpawnPointManager to actually store the Spawn Points and perform opertations on the Spawn Points.