NetworkSpawnPointView now spawns multiple Energy types based on clicks
NetworkSpawnPointView now checks for how many times a player clicks up to 3 times and sends that to the PlayerInput to allow the player to choose how strong the energy will be.
NetworkPlayerInput now passes in the energyCost as the energyIndex to state on what energy type it will spawn.
NetworkEnergySpawner now use the energyIndex to grab energy data rather than hardcode a standard energy.