Added TestEnergyLaoder and hooked it into the NetworkEnergySpawner
NetworkTestEnergyLoader creates energies in the pattern of the lower the cost, the weaker it is and the faster it is. Inversely the more expensive it is, the stronger it is and the slower it is.
Network main connects the NetworkTestEnergyLoader to the NetworkEnergySpawner and loads the energies when the game starts.