Added RoundManager to GameLogic
Created associated unit tests for RoundManager.
RoundManager stores the number of rounds for a match and calculates the number of rounds needed to be won in order to win a match. It also stores the id of a winner with the associated round they have won and calculates when a unique player id has won enough rounds to win the match.