Created RoundManagerVisual interface for RoundManager to inform when the visual should start and when it has ended
IRoundManagerVisual forces implementation of starting the visual of a round ending and countdown to start of the next one as well as informs RoundManager that the visual is done and the next round can proceed.
Created NullRoundManagerVisual script for testing purposes.
Created RoundManagerPlayModeTest to test the ability of RoundManager not giving the RoundVisualISDone go ahead until the visual says.