Created GameEnd and Rematch Functionality
NetworkMain now creates and spawns the rematch screen on the server then connects it to itself.
NetworkMain now has Rematch functionality which creates a new RoundManager, connects the visual to it and sets the new round manager for the GameManager before starting the game.
NetworkPlayerInput has a UNET specific work around to allow the rematch screen to recieve client authority as to be able to send RPC's and Commands.